opea-comps/megaservice/requirements.txt
opea-comps
httpx
